fix: avoid possible value overflows

+++ src/Ray.BiliBiliTool.Config/Options/DailyTaskOptions.cs
@@ -80,12 +80,12 @@ namespace Ray.BiliBiliTool.Config.Options
         /// <summary>
         /// 自定义漫画阅读 comic_id
         /// </summary>
-        public int CustomComicId { get; set; } = 27355;
+        public long CustomComicId { get; set; } = 27355;
 
         /// <summary>
         /// 自定义漫画阅读 ep_id
         /// </summary>
-        public int CustomEpId { get; set; } = 381662;
+        public long CustomEpId { get; set; } = 381662;
 
         public List<long> SupportUpIdList
         {
